# Conflicts:
#	libzkbob-rs-wasm/Cargo.toml
#	libzkbob-rs-wasm/src/client/tx_parser.rs
#	libzkbob-rs-wasm/src/ts_types.rs
#	libzkbob-rs/Cargo.toml

+++ libzkbob-rs/src/sparse_array.rs
@@ -129,7 +129,7 @@ where
         self.db
             .iter(0)
             .for_each(|(key, _)| {
-                batch.delete(0 as u32, &key); 
+                batch.delete(0_u32, &key);
             });
         self.db.write(batch).unwrap();
     }
